requirements.txt
Flask==2.1.2
Pillow==9.4.0
numpy<2.0.0
tensorflow